Restructure project files
The project files should look a lot more similar to how you would see a
web development project hopefully. The goal is to try to make game dev
more accessible to web developers.
