+++ src/core/character/interface.ts
