+++ src/core/party.ts
@@ -1,4 +1,4 @@
-import CombatableCharacter from './characters/combatable-character';
+import { CombatableCharacter } from './character';
 
 /**
  * A class for combining combatable characters together to create a party
