+++ src/scenes/battle/components/character-control.ts
