+++ src/scenes/battle/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CharacterControl } from './character-control';
+export { default as BattleControl } from './battle-control';
