--- src/scenes/components/battle-controls/index.ts
@@ -1,1 +1,0 @@
-export { default as CharacterControl } from './character-control';
