+++ src/scenes/components/index.ts
@@ -1,2 +1,1 @@
-export { default as BattleControl } from './battle-control';
-export { default as CharacterSprite } from './character-sprite';
+export * from './sprite';
