+++ src/scenes/components/sprite/index.ts
@@ -1,0 +1,2 @@
+export { default as CharacterSprite } from './character-sprite';
+export { default as PartySprite } from './party-sprite';
