+++ src/scenes/index.ts
@@ -1,0 +1,2 @@
+export { default as BootScene } from './boot';
+export { default as BattleScene } from './battle';
