+++ test/character/hero.spec.ts
@@ -1,4 +1,4 @@
-import Hero from '../../src/characters/hero';
+import Hero from '../../src/core/character/hero';
 import { expect } from 'chai';
 
 describe('creating a new hero', () => {
