+++ test/character/monster.spec.ts
@@ -1,4 +1,4 @@
-import Monster from '../../src/characters/monster';
+import Monster from '../../src/core/character/monster';
 import { expect } from 'chai';
 
 describe('creating a new monster', () => {
